Created a spawner for obstacles below
Calibrated the speed for each obstacle so that they are created at the same time and do not interfere with each other.  Reduced the speed a bit.  In order for the obstacle to be created from below as well, instead of creating copies of the three previous ones and loading the scene once again, I will transform the position of new random obstacles directly in the spawner in the Spawn function by creating a new random counter.  More details can be found in the comment file itself.